Переносит константу в const.js
